requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
selenium==4.11.2
pandas==1.5.3
numpy==1.24.3
yfinance==0.2.28
ta-lib==0.4.28
python-dotenv==1.0.0
sqlalchemy==1.4.41
pydantic==1.10.12
webdriver-manager==4.0.1